requirements.txt
azure-ai-textanalytics>=5.3.0
python-dotenv>=1.2.1
pyyaml>=6.0.3
